feat: separate web editor into own package
