+++ packages/harper-editor/src/app.d.ts
@@ -1,0 +1,5 @@
+declare global {
+	namespace App {}
+}
+
+export {};
