+++ packages/harper-editor/src/lib/index.ts
@@ -1,0 +1,4 @@
+export { default as Editor } from './Editor.svelte';
+export { default as LazyEditor } from './LazyEditor.svelte';
+export { default as LintCard } from './LintCard.svelte';
+export { default as LintSidebar } from './LintSidebar.svelte';
