+++ packages/harper-editor/src/lib/styles.css
@@ -1,0 +1,5 @@
+@import "tailwindcss";
+
+@custom-variant dark (&:where(.dark, .dark *));
+
+@source "./**/*.{svelte,ts}";
