+++ packages/web/package.json
@@ -38,12 +38,12 @@
 		"components": "workspace:*",
 		"drizzle-orm": "^0.44.6",
 		"drizzle-zod": "^0.8.3",
+		"harper-editor": "workspace:*",
 		"harper.js": "workspace:*",
 		"lint-framework": "workspace:*",
 		"lodash-es": "^4.17.21",
 		"mysql2": "^3.15.2",
 		"posthog-js": "^1.245.1",
-		"quill": "^2.0.3",
 		"reveal.js": "^5.1.0",
 		"svelte-ace": "^1.0.21",
 		"svelte-intersection-observer": "^1.0.0",
