+++ packages/web/src/routes/editor/+page.svelte
@@ -1,8 +1,8 @@
 <script>
 /// This page exists to be embedded via an `iframe`.
 
+import { Editor } from 'harper-editor';
 import { page } from '$app/stores';
-import Editor from '$lib/components/Editor.svelte';
 import Isolate from '$lib/components/Isolate.svelte';
 
 let content = $page.url.searchParams.get('initialText') ?? '';
